Store shop configuration also in an include()-able file.
This is temporary code to get the migration started. The file gets
written on changes to existing URLs, only.

Thie idea is to store shop configuration not inside the database,
but in a regular file, to make it accessible for upload scripts.
This should ease cloning or synchronizing shop a lot.

The approach taken is to use a global array, which is written to
a file, which in turn can be loaded by an include(). Nice side
effect should be better performance, because this rarely changing
data can be cached as PHP code.

Next steps are to change accesses to this data to the new global
variable, then to remove the inheritance from ObjectModel, which
also makes DB storage obsolete, then.