AdminController: introduce comparator for sorting DB-like arrays.
So far this was done with ORDER BY clauses in SQL queries. When
reading data from files we no longer have such queries, so we
need another way for doing such sorts.

Code is part of class AdminController, because it depends on
properies of that class.

Typical usage (for AdminController derivates):

  $this->_orderBy = $orderBy;
  $this->_orderWay = strtoupper($orderWay);
  usort($result, array('self', 'compareByArrayValues'));