Introduce ObjectFileModel for file based storage.
So far just enough to load ShopUrl objects from the file based
storage and a proper write function. For example, after adding an
URL one has to "edit" that URL to trigger a write in class
ShopUrl.

The cache clearing stuff was added later. Some functionality
triggers immediate redirects after the storage part is done, so
having the refreshed file available immediately is pretty crucial.

- - -

The idea here is to provide an alternative to class ObjectModel,
which stores data not in the database, but in a regular file. This
has at least two advantages:

 - Chance to tweak parameters by external scripts with reasonable
   efforts. Very interesting for these domain and URL settings,
   because these change by just cloning a shop to another location.

 - Better performance, especially for rarely changed data. Because
   data can be read on the PHP code level, with a simple
   include_once() statement. This is in turn cached by the code
   cache, so performance should be as fast as one can go with PHP.

Implementation is to make ObjectFileModel a shim between the
calling class and ObjectModel, just overriding all the methods
doing database accesses (there are many methods not touching the
DB at all).