ObjectFileModel: get rid of the global variable.
Finally! This is the fourth test implementation attempting to
eliminate this nasty global. The solution ist to strictly use
accessors, only. Even from inside the class. This way a method
can mimick a class variable and as it's a method, an include() is
allowed :-)

While include()'s can't deal with changed file content during a
run, it's possible to switch to a class variable, then. See
'protected static $storage' in ObjectFileModel. This code was
commented out, however, because it isn't needed for subclass
ShopUrl. Still kept for future reference.

Also worth a note: at install time, config/shop.inc.php doesn't
exist, so one has to test for the existence of $storage before
using it after the include().

Note to developers: if you had a config/shop.inc.php before,
change the variable name from $configUrl to $settings inside this
file. As storage is fully encapsulated now, code uses the same
variable name for all subclasses (but distinct storage files).

Fix ObjectFileModel.

This commit should be used to fix the original commit.

I was pretty sure to have tested wether is_array() works with
an undefined variable. It turns out it doesn't and throws
an error message, instead.