Carrier wizard: don't mess with global stuff when checking zones.
It's a bit unclear what the point of the previous code was. Perhaps
some good, but fancy idea. Nevertheless, configuring carriers is a
task done rarely, so it should be manageable without taking
lessons.

Now the value entry field enables and disables the standard way.
Additionally, empty fields get populated with zeros when enabled
and zero number get removed when disabled. This is hopefully
helpful for intuition.