Carrier wizard: remove concept of 'valid' ranges.
This concept applied at page edit time, only, so no change in
what's stored permanently.

I think the idea was as following:

 1. Click checkboxes for the wanted zones.

 2. Make the/a range valid by entering boundaries.

 3. See how the price fields get enabled magically as soon as
    the boundaries are valid.

The unfortunate thing here is, this also disallowed editing prices
as long as the related range wasn't valid. Which is bad, because
it confused users. Validating just before saving only is much less
confusing.