Carrier wizard: remove semi-automatic range-rearranging.
Not sure who had the idea of such code. If I hadn't looked into
the source code I likely never had found this. It's hidden by
default and it's unclear on how to unhide it. After unhiding with
Firefox' web developer tools I couldn't get it into doing
something useful.

It's probable purpose is to re-order ranges which were entered in
the wrong order, but are fine in every other aspect. Kind of an
attempt to semi-automatically fix just one of hundreds of possible
user mistakes. That's pointless, especially if it requires some
100 lines of complex code.