Introduce tools/generatemd5list.php.
That's a PHP version of tools/generatemd5list.js. The JavaScript
version was suspect to create subtle differences.

In a develoment installation, output of the JS version is the same,
except for order/sorting and some files directly in admin/ missing.
While building a package however, the JS version picks up some 900
additional files, some of which can't be parsed by the back office
verifier.