Theme17: remove 1.6 vars for 1.7 themes, boilerplate page.
This is for the boilerplate page when calling a product page link.
The actual product page and other pages are about to follow.

As an unfortunate fact, the set of variables passed to a 30bz/
PS 1.6 compatible theme is entirely different from that passed to
a PS 1.7 compatible one. Accordingly, a good first step towards
dealing with PS 1.7 compatible themes is to get rid of unwanted
variables, first.

Smarty assignments missing for the PS 1.7 theme will be added by
the following commits, step by step.