INCOMPLETE Theme17: add trait FrontController17.
TODO: remove all this commented out code.

Also restore a couple of simple Smarty variables.

General note: code for this trait in this and the following
commits was copied straight from FrontController.php as of 1.7.0.0,
then reviewed and tested for equivalence of the Smarty variable
set on a single product page.