INCOMPLETE Theme17: enable Smarty variable $logged for PS 1.7 themes.
TODO: find out why this is needed in theme jump-ing, but not in
      classic.

For some unknown reason this is needed for the RepRap DIY
modification of the Classic theme, file ps_customersignin.tpl.
The unmodified theme uses the same variable in the same place,
but doesn't trigger a Smarty error.