InformationController: re-establish test error messages.
Code is a bit more complex than necessary, but this way we get
away without changing the template. Which is crucial for those
people using a custom backoffice template.

A simpler solution for both, PHP code and template, whould be to
forward an array with only errors for each, required and optional
tests. Array not empty -> an error exists, then list all array
entries, case done.