FileLogger: don't die on failures.
Dieing has the obvious disadvantage that code after an attempt to
log a message isn't executed. Having no log message is
disappointing, but by no means fatal.

This should also solve the problem of the full log file path
( = server internals) being exposed to customers on logging
failures, which was reported as issue #367.