PageNotFoundController: don't depend on $_SERVER['REDIRECT_URL'].
New solution is to try to match both, redirected and not
redirected URL (makes a distinction with Friendly URLs enabled).
This shouldn't be a performance issue, as such requests no
longer happen once the image got created.

This should solve issue #474 ...

... but exposes a couple of other issues, like product images in
size 'home_default_smaller' requested, but not being rebuilt.

Further, images img/c/3-1_thumb.jpg and img/c/3-0_thumb.jpg are
committed directly into the repo instead of being in the set of
demo data and neither matching nor being rebuilt when requested.

Last not least, there are a few nore places which depend on
$_SERVER['REDIRECT_URL'].