validatemodule.sh: test wether the module needs a release.
This is a fairly complex operation, because these two more trivial
approaches are considered to be insufficient:

First trivial approach is to expect 'master' and the latest
release tag to be identical. While this would work (remember, we
test this only when doing releases anyways), it would also enforce
making a release for insignificant changes, like e.g. improvements
in the buildscript. Eventually we could have two releases with
identical content.

Second trivial approach is to simply not test the distinction
between 'master' and the latest release tag. However, this allows
to simply forget making a module release, which has happened
already.

The now choosen approach is to do an automated best guess on
whether changes are significant enough to require a release:
if packaged files got changed, a new release is needed.