validatemodule.sh: replace ${LS} with ${FIND}.
'git ls-files' works on the index, 'git ls-tree' on a revision.
While this is effectively the same when querying the currently
checked out revision, this change is one step towards validating
arbitrary revisions. It's also a step towards validating non-
repositories, because output of 'git ls-tree' is closer to that
of the non-git 'find'.

Some code places needed additional adjustments, 'git ls-tree'
doesn't support wildcards. All code changes carefully compared
to the previous effect, of course.