build.sh: become independent from HEAD.
Files get pulled from the repository in the requested version
without a need to check that out. Which means, it no longer
matters wether that repo is dirty or not.

This is a major step towards making stashing the current
repository state unnecessary. And it took quite some research to
find the commands needed for this, because 'git submodule' works
only with the current working copy.

Essentially, the strategy is to use Git plumbing commands to
traverse the tree of a given revision manually. Not too much
effort, because thirty bees has its submodules in only two
directories, modules/ and themes/. Current code relies on this
fact.

Should ever appear more directories with submodules, this code
will find all of them:

  for M in $(git cat-file -p ${GIT_REVISION}:.gitmodules | \
               grep '^\[submodule' | \
               cut -d '"' -f 2); do
    echo "${M%/*}"
  done | sort -u

Packaging result verified to be identical to that before starting
the recent build.sh effort.