build.sh: make sure all required submodules exist.
All of the submodules required for the currently checked out
revision exist, of course, but if another revision gets requested
that other revision might have a different set of submodules.
Accordingly, clone them as needed. And remove them after being
done, of course.

This is the second and last part to become independent from
checking out the requested revision. Now everything gets pulled
out of the repository without even touching the current worktree.

Packaging result verified to be identical to that before starting
the recent build.sh effort.