Class Module: don't put HTML comments around JSON.
Modules can return JSON. HTML comments invalidate this JSON.

Noticed case was adding a product to the cart with debug mode
turned on. Wouldn't work, because the Ajax call for doing so
returned just Null (instead of JSON with cart contents), resulting
in the 'Add to cart' button dimming and waiting, but no cart
content dialog popping up.

The bug appeared with commit 1a8f9a61, between 1.0.4 and 1.0.5.