build.sh: KEEP paths shall start at repository root.
This allows to keep directories at repository root, while still
excluding like-named directories deeper in the file tree.

Change is rather big, because there's also a move towards using
'find' with '-prune'. Because it was found to work more reliably.
And yes, '-print' makes a distinction and '-delete' isn't
compatible with '-prune'. 'find' can be surprising :-)

Packaging result verfified to be unchanged.