build.sh: package only modules going to get actually installed.
We have a list of these modules reachable for scripting now, so
take advantage of it.

This allows to add more submodules to the core Git repository and
exposing them to Travis CI tests without increasing the size of
the release package. It also reduces the disk footprint of a
default shop installation.

Merchants will barely recognize the distinction, because the
module list in back office will still list all available modules,
no matter whether they are on disk or available online.