Add paragonie/random_compat package to dependencies
Rijndael depends on `random_bytes` function, which is defined in php7
only. For php5 we need to use polyfill. Package `paragonie/random_compat`
provides polyfill for this function. This package is already included
because it's dependency of `defuse/php-encryption`. This commit
adds direct dependency, and also ensures that the package
version is not 9.99.99 - see issue #593