Remove back office tabs no longer used in thirty bees.
These are ones which existed in PrestaShop.

Putting this into the general upgrade procedure, because the
migration module didn't do this so far, so all the migrated shops
still have these tabs. It's a no-op for shops which initially
started with thirty bees or removed this tab manually already.

Related to issue thirtybees/psonesixmigrator#15.