1.0.8 upgrade: add back office tabs new in thirty bees.
These are ones which didn't exist in PrestaShop.

Putting this into the general upgrade procedure, because the
migration module didn't do this so far, so all the migrated shops
still miss these tabs. It's a no-op for shops which initially
started with thirty bees or added these tabs manually already.

Related to issue thirtybees/psonesixmigrator#15.