Cart: recalculate taxes for totals as needed.
This should fix remaining parts of issue #795.

The option to kind of re-engineer the tax rate by comparing a
price with and without tax was choosen, because applying taxes is
a pretty complex business and the last thing we need is another
place implementing it. Computation precision should be sufficient.