ajax.php: fix the query for AdminReferrersController.
Commit e94801061e559942c767d140b2d6703a6c6bfcb5 (between 1.0.0
and 1.0.1) moved this class-less code in AdminReferrersController
from outside the class into the constructor, so it was no longer
executed by a require().

Solution is to move the code back into ajax.php, where it was
years ago, to not introduce the unwanted concept of class-less
code in files for classes.