Mail::Send - do not include lang.php with email subjects
This method tries to include lang.php file, which contains $_LANGMAIL global
variable with translated email subjects. But this global variable is not used
anywhere in this method. The only place in the thirtybees core where $_LANGMAIL
is used is Mail::l() method. And this method includes the lang.php file itself.

Moreover, Mail::Send() method expects that subject is already translated.
This is somewhat enforced by translation tool, which looks for code like
this:

Mail::Send($lang, 'template_name', Mail::l('subject', $lang), ...)

When thirtybees finds (using regexp) this code in php files, it will create
association between template 'template_name' and subject 'subject', and
offers this email subject string for translation. The important part is that
it looks for Mail::l inside Mail::Send method, which means that $_LANGMAIL
global variable is already loaded when Mail::Send method is called.

This means that including lang.php file inside Mail::Send is unnecessary,
and can be safely removed. Also, this fixes the issues with emails not
being send when lang.php file is missing for given language in /mails/
directory

Related to issue #871