Installer: raise minimum memory requirement to 128 MiB.
This is quite a tep upwards from the previous 32 MiB. Reasons
include:

 - A (rather conservative) default PHP installation on
   Debian/Ubuntu meets this requirement, so everybody should have
   at least this.

 - Even the installer tries to raise this limit to 64 MiB already
   in order to work reliable. However, such limit raises from
   within PHP code can fail. Relying on such raises anyways is
   just asking for trouble.

 - Many modules and some classes try to raise this limit as well.
   With the same set of implications.

 - It's just an upper limit. Actual memory usage is lower.

 - For comparison: the testsuite, which loads all classes,
   requires some 1080 MiB of memory on PHP 5.6, so guaranteeing
   10% of this is certainly not impudent.

Plan is to remove all attempts of raising the memory limit. Cases
not working with 128 MiB either have to get split into smaller
chunks (preferred) or report an error if the limit isn't
sufficient (should be a rare need). To get reliable operations.