build.sh: move 'continue' out of the subshell.
Apparently newest Bash (5.0.3) has changed its mind about behavior
of a 'continue' in a subshell. As a result of this, build.sh
started to ignore the 'continue' and adding non-default modules to
the release package (and reporting a warning). Message was:

  continue: only meaningful in a `for', `while', or `until' loop

This commit restores previous behavior on newest Bash and should
be compatible with older Bash versions.