Shop: fix loading (multi)shops with a virtual URI.
This fixes a regression introduced with be67c7f936ee89352a41,
first hunk.

Shop URLs need to be ordered by length, else a shorter URL
contained in a longer URL is found first. Examples:

A DB query result of

  [...]
  'uri' => '/thirtybees/',
  [...]
  'uri' => '/thirtybees/two/',
  [...]

finds the shop at '/thirtybees/' in a request URI '/thirtybees/two/',
while a query result of

  [...]
  'uri' => '/thirtybees/two/',
  [...]
  'uri' => '/thirtybees/',
  [...]

correctly finds the longer match.

Maybe that's still not bullet proof, but that's how PS did it.

This might fix issue #912.