Succed  a search even with spelling errors
Hi,
Here, just an implementation of the levenshtein algorythm.
https://www.php.net/manual/en/function.levenshtein.php

The goal is to find some products with the search functionality. Even if you get some miss spelling in the search input.
That works perfectly with the vanilla ThirtyBees shop with default product.

For example, you search  "chonney" instead of "honey", search controller will be able to find a close work, here honey, and will display honey product in the result.