Tests: Fix BasicGuestFiveStepCheckoutCept test
Niara theme hides native checkbox <input> on checkout page, and replace
it with svg version.  Unfortunatelly, this break the acceptance test,
because the element is not not visible.

Fixed by changing the test logic - instead of setting <input> value, we
simply click on its label. That will, in turn, toggle associated <input>
value.