Translate::getModuleTranslation - fix escaping bug
Output of Translate::getModuleTranslation used to be different when
translation php file existed or not.  For example, following translation
string:

$this->l('There\'s bug');

would return
   There's bug   -- in case en.php file exists
   There\'s bug  -- if en.php file doesn't exist

The reason for this behaviour is that Translate::getModuleTranslation
used different output escaping for both branches.

The fix is to unify both branches to use one output escaping method.