Product: fix duplicate functionality
After product or combination saved, thirtybees calls method
SpecificPriceRule::applyAllRules() for this product. This method
creates specific prices according to existing rules.

This does not work correctly with duplication functionality, because
some specific prices can have duplicate unique key.

When duplicate product, we don't really care about specific prices
generated by rules, because we want the same rules as defined in
source product. Therefore, the fix is to delete any rule that was
created during product / combination creation.

It would be better to not create these specific prices in the first
place, but unfortunately we don't have enough information to make this
kind of decision in add() method.

Closes #893