ConfigurationTest: change test for TLS v1.2 connectivity.
As https://tlstest.paypal.com/ apparently ceased to exist, we need
a distinct method of testing for this kind of connectivity. Using
some kind of hobby service was discarded for its unknown
reliability and because such usage could be considered unfair.

New solution is to use https://api.paypal.com directly. One can
distinguish a connection failure from an authentication failure,
mission accomplished.

This is related to issue #1007.