ASM: fix bug with missing id_warehouse in order
When merchant used added new ASM product with combinations into the order,
and didn't changed default combination, then the product was added with
id_warehouse == 0.

This was because the javascript uses strict comparison to find default
combination:

  if (this.default_on === 1) {
    ....
  }

This strict comparison failed because default_on property was passed
from php as string "1" or "0", instead of int.

Closes #373