AminController: improve module enabled check
When AdminController collects information about tabs that should be
displayed in back office menu, it has to perform enabled check. In order
to do so, it used to instantiate module, and call instance method

$module->isEnabledForShopContext

This means that the module classes fully loaded, and constructor called,
even for modules that are actually disabled. This is not optimal from
both performance and security point of view.

This commit changes the situation by extracting the mmodule enabled
check to static method, so module does not need to be instantiated.