AdminGeolocatinController: select service
There are many geolocation service providers, or ways to perform
geolocation. Thirtybees always used maxmind database to implement
feature this in the core. Choosing service provider should not be
decision of thirtybees, however.

This commits allows merchants to select golocation service provider they
want. It depends on hook 'actionGeoLocation' -- any module that
implements this hook will be offered for selection