Remove geoip library dependency
Thirtybees core no longer use maxmind geoip library in the core to
perform geolocation. The responsibility to do this was moved to service
modules. Therefore, it's no longer useful to have this php library as
part of thirtybees distribution.

This is compatiblity change, of course. Any module depending on this
library will no longer work. However, this should not be a big deal,
since the library v1 is severy outdated and no longer maintained.
Also, database for this version is no longer available for download from
maxmind website.

Even though I do not expect this change to have big impact, there can
be some merchants that will experience problems with this library
missing. For these, there will be compatibility module available.