Image::deleteImage - don't fail on missing image
The purpose of deleteImage is to ensure that image files are removed
from the filesystem. Right now, the method returns false if the primary
file is missing. This also means that thubmnails and other derived
images are NOT removed from the filesystem.

The better way is to continue the delete process even if primary file is
missing.