Configuration::updateValue: preserve new lines
There was a bug introduced in 1.1.0 that resulted in new lines being
striped away from configuration values, if $html escaping was not used.
This bug was caused by this code:

  strip_tags(Tools::nl2br($value));

Which is obviously wrong. nl2br converts new lines to <br />, and
strip_tags immediately removes this tag from the code.

The fix is to remove nl2br call altogether.

Closes #1107