Dispatcher: introduce custom matcher functions
This commit adds a new option to define custom matcher function for
dispatcher routes. The intent of these functions is to perform additional
checks on matched URI, and decide if this URI indeed belongs to the
corresponding route.

For rewritable routes (product, category, this check functionality was
already present. This commit mostly extracts and generalize this idea.

By extracting the check logic, and associating it with specific routes
only, this commit also fixed issue #1108.

Closes #1108