Supply Orders: manual validation for details
ObjectModel::validateController() not only validates input values, but
also copies data from request to current object properties.

This means that we should call this method only for *primary object*
associated with controller, in this case SupplyOrder. We should not call
it for SupplyOrderDetail detail entries, because that result in storing
SupplyOrder data inside SupplyOrderDetail entries.

Originally, the method validateController was overwritten in
SupplyOrderDetail. This override was removed in commit
078765ff2c1f8f81501b076890f353efb60e79d8, and that introduced
issue #1035.

It's not necessary to restore this override, becase most of the
validation it implemented is also performed during the object model
save() call.

The only things that are not checked upper/lower are limits for
quantity_expected, tax_rate, and discount_rate fields.

Closes issue #1035