displayPrice: always replace USD symbol
When Auto format is turned on for some currency, displayPrice javascript
method will use browser built-in support for currency formatting to
format currency in specific locale.

It does so by evaluating this expression:

return price
  .toLocaleString(locale, {
    style: 'currency',
    currency: 'USD',
    currencyDisplay: 'code'
  })

This code ensures that the number is formatted using 'USD' string in
place of currency symbol. This 'USD' string is then replaced with actual
currency symbol.

However, there was a bug - this string replacement wasn't performed when
no currencySign was passed to the function. In such case, the string
with 'USD' symbol was returned.

This commit ensures that this USD symbol is always replaced, even with
the empty string.

Related to issue #865