Guest checkout: relax email validation
When guest checkout is enabled, it's possible to use same email address
to create multiple guest accounts.

However, if customer account with the same email address already exists,
guest account creation is not allowed (not even one)

This commit relaxes this restriction -- when guest checkout is enabled,
it's always possible to checkout, even if customer account already
exists.

Closes #1150