Order: fix adjusting of total_paid_real
This commit fixes problems with adjusting `total_paid_real` property
value. In some use cases, the input amount was not converted to order
currency, resulting in invalid value being saved in this column.

Note that this column/property is deprecated, and should not be used anywhere
in core. It exists for compatibility reasons only.

The value of this column should always match sum of all order payments
from `tb_order_payment` table converted to order currency.

Closes #1161