Cart::getPackageList - fix ASM bug
When ASM is enabled, and ASM-enabled product have no warehouse
association, then Cart::getPackageList() returned package with no
warehouse association, but also with no available carriers.

This is unwanted behavior - because product is not associated with
any warehouse, default carriers should be returned. The result should be
the same as in the case of non-ASM product.