Installation: display uncaught exceptions
All unexpected PrestaShopException exceptions should be displayed to
users in debug mode.

Displaying standard error page is not the right solution here, because
this error page has only an option to download encrypted exception.
Because installation failed, there is way to log in to administration,
and decrypt this exception file.

This commit also solves the situation when the exception is thrown
during ajax call. In that case, error message is serialized to standard
error json payload.