Fix WebserviceOutputBuilder::renderAssociations
Webservice expects $objectsAssoc to be array or arrays, hoverer some ws
methods return array of ids.

The proper way to fix this would be to modify these ws methods, however
that would be compatibility change -- these methods are unfortunately
used by third party modules.

This commit fixes the problem by allowing array of ids as well as array
of arrays.

Closes #1219