Profile::getProfileAccess - always return array
Method Profile::getProfileAccess could return false in case Tab does not
exists in database. The rest of the code in the core does not expect
this, it always assume valid permission array is returned.

This commit fixes this problem by always returning permission array,
even if Tab does not exists.