BO order creation: check product quantity early
When creating order in back office, it is possible to set custom price
for products. System creates temporary specific price for them - this
specific price should be deleted after the order is created.

However, only after the specific price is created, system checks product
quantity on hand. If there is not enough quantity, the order process
ends with error, and specific price remains in the system.

This problem is fixed by this commit, simply by reordering these two
steps. We first check product quantitty, and create specific price only
after we are sure we can continue.

Related to #1343