Category::getName - optimize
Small performance improvements. If the category object was loaded in the
same language context for which getName() method is being called, we can
directly return value stored in $this->name field.

If getName is called for different language, we still need to retrieve
the correct information from database.

Related to #1277