StockManager: stock movement fails in FO context
In some scenarios, ASM stock movements can be initiated by front office
actions.

However, code in StockManager always expected these operations to be
triggered from back office. Because of this expectations, the code tried
to access Employee object stored in Context. In case of front office,
this variable is null.

Related to #1178