Link: image link without watermark
When watermark module is installed, merchants can select image types
that should be protected with watermark image.

There is also an option to display original (unprotected) image for
logged-in customers.

When this option was selected, thirty bees failed to generate correct
link url for those image types that were excluded/not selected for
watermarking.

This commit fixes the issue at hand. It's not the best solution, mainly
because this functionality should not be in the core to begin with. We
shold introduce some imageType postprocessing action hook instead, and let
watermark module implement this hook properly.

Closes #1356