Deprecate function rewriteSettingsFile
Function rewriteSettingsFile was supposed to regenerate settings.inc.php
file. This function is severely outdated, and generates incorrect
result. Some new settings keys, for example _TB_REVISION_ or
_PHP_ENCRYPTION_KEY_, were omitted from settings content.

The only usage of this function in thirty bees codebase was in media
server configuration handler. Calling this function in this context does
not makes any sense, as media servers are not saved inside
settings.inc.php file.

This commit converts this function to noop. The function is candidate
for removal in future releases