Cookie: don't save empty cookies
Cookie contains initialization logic to force language detection that
marks cookie as dirty, even if it's otherwise empty. This leads to
situation when cookie is always created and send to client, even if it's
empty.

For example, consider following simple code:

$cookie = new Cookie('psAdmin');
if ((int)$cookie->id_employee) {
   ...
}

Such code is often used by modules to check if current user is logged-in
to back office (administrator). If the user is not logged-in, then the
new cookie is created, and send to client. This is not desired behaviour.