Webservice: fix product categories association
Commit ebf6f91131a7adf43a761e39cf51f8587a549a8a introduced a new bug
that reproduces when dissociating all categories from product.

Method Product::updateCategories([]) returns false when called with
empty parameter, and that in turn cause webservice call to throw an
error.

The fix is simple - if no categories are to be associated with the
product, then simply remove product from all categories and finish