Dashboard: auto recalculate date range
Fixed logic to auto-recalculate date range on dashboard. Once a day, the
start and end dates for currently selected date range preset will be
recalculated, and stored inside employee record.

This mechanism uses cookie to determine when to recalculate dates. This
allows for filtering on custom date range period - start and date range
will not be recalulated until midnight.

This is not the best mechanism, and it should be replaced with explicit
'custom' date range preset. Tracked as an enhancement request #1361.