Category: properly save groups
There was a problem is with category group handling.

When category was created or updated by AdminCategoriesController,
property $groupBox was populated properly, and groups are saved to
database.

However, when category was created or updated programatically, and
property $groupBox was not set, then groups were initialized to default values

Code example:

	$category = new Category(123);
	print_r($category->getGroups()); // result = [1]
	$category->active = true;
	$category->save();
	print_r($category->getGroups()); // result = [1, 2, 3]

The fix modifies the behaviour of $category->groupBox property. If it is
not explicitly set to array, then add/update function will ignore it,
and group list will not be changed.

Closes #1363