Dispatcher: fix keyword 'param' support
Dispatcher routes can specify 'param' for keywords used in route url.

Dispatcher will extract the parameters from url, and inject them into
$_GET array under the 'param' name.

However, when generating links, the dispatcher used 'keyword' name
instead of 'param' name. Everything worked correctly, as long as 'param'
and 'keyword' name were the same. While this is true for most cases,
sometimes these two identifier are different, leading to weird errors.

This commit fixes this issue by expecting 'param' name when generating
links. For backwards compatiblity reasons, if the 'param' is not
provided, but 'keyword' is, it will fallback to this value.