Do not ignore SQL errors by default
Unless debug mode was enabled, thirty bees always silently ignored errors
raised by database when executing SQL queries.

First problem with this approach was that execution path was different
for debug and normal environments, which actually make it harder to
debug.

More imporantly, database errors should never be ignored. This usually
causes other errors down the line. These subsequent errors are very hard
to debug and understand. It's much better for everyone when database
exception is thrown early.

This is a compatibility breaking change, of course.

For backwards compatibility reasons, a new switch was introduced in
Performance tab. This switch allow merchants to ignore sql errors.
Hopefully this will be used rarely.