Product search: abort previous ajax requests
A lot of merchants add orders from the BO. There is a product search
functionality which with ajax, but doesn't work flawlessly.

The search is triggered while typing without a big delay time. If you
type fast (as devs do), there is no issue. The ajax call is only made
once. But if you type a bit slower, multiple ajax calls are sent. It's
basically by chance, which result is rendered. It's just the last one,
that comes back. It often leads to bad results.

The fix is to abort previous ajax requests that have not finished yet.