HelperList: support custom drill down url
When HelperList renders list onclick handler is generated for each cell.
This onclick handler opens the record in view display mode.

In some cases, it would be usefull to have custom link, for example link
to different controller. Or link to different action (edit instead of
view)

To facilitate this, we are introducing new HelperList property
$linkUrlCallback. If this property is set to valid callable, it will
be called for every row. This callable is expected to return valid url,
or null if the row is not clickable.