Dynamic Packs feature
Implementation of dynamic packs functionality. Users can finally create
pack products for which quantities are automatically calculated based on
quantities of its items.

While thirty bees core makes sure that pack quantities are in sync with
items quantities, some modules (or even some hidden part of the core)
can modify product quantities in a way that bypass synchronization
process. If that happens, pack quantities can be out of sync for a
while. To limit this situation, new work queue task was created, and is
scheduled to run every 8 hours.

Closes #495