PaymentModule: fix customer context initialization
Method validateOrder always initialized context $customer variable to
match the customer associated with the cart.

This operation should be performed only when necessary - if the context
is not initialized yet for this customer.

In most cases, payments are processed synchronously as part of the
checkout flow. PHP code is executed as part of the request handler that
was triggered by browser, and cookie was already used to initialize
context properly.

Sometimes, payments can be processed asynchronously. For example, payment
gateway can trigger server to server call to notify thirty bees about
payment status. This webrequest does not contain customer cookie, and
therefore we have to initialize context properly.

Closes #1331