Fix methods with optional params before required
Optional parameters should not appear before the required parameters in
a function/method declarations. Such method have to be called with all
required parameters anyway, so it does not make any sense.

This issue also triggers PHP warnings in PHP8

The only possible problem here is with overrides. However, changes that
any of these methods were overriden are small.