Fix php warnings in Tools::getBrightness
Method Tools::getBrightness should be called with valid color in hex
format, such as #ff0010. However, there are places in core / modules that
can pass color name instead, such as 'red' or 'blue'. With such inputs,
the method raised PHP warnings, and returned 0.

The fix is to verify that input value is a valid hex color. Also, some
well known color names are supported, as a small bandaid.