Remove deprecated auto_detect_line_endings
PHP directive auto_detect_line_endings is deprecated in PHP 8.1 and will
be removed in PHP 9.

Removing this code should not have any impact. This directive only
enables line-ending detection for \r characters, which were only used in
legacy Mac OS environments. It is very unlikely that anyone in 2022
would try to import file that is using \r as a line separator.

For more information, see
https://php.watch/versions/8.1/auto_detect_line_endings-ini-deprecated