Add/remove override: fix evaluation
When adding or removing overrides, thirty bees have to evaluate slightly
modified override file content in order to determine class methods,
fields and constants using PHP reflection.

One of the modifications removes `extends` clause. This modification
exists to prevent loading of core classes, as it is not needed.

Unfortunately, in PHP8, this chokes if there is call to parent method
using syntax `parent::method()`. Even though the class is not
instantiated or method is called, PHP8 raises excetion during initial
class parsing phase. In PHP7 the exception would be called during
runtime, so this was not an issue.

This commit fixes this problem by making all evaluated overrides extend
\stdClass class. PHP8 parser is happy, even though `parent::method` does
not exists.

In the future, the whole evaluation/reflection mechanism will be
replaced by php parser to make the whole process much more safer and
stable.

Closes #1470